Fixed getTotals()
https://github.com/opencart/opencart/pull/14903

+++ upload/catalog/model/checkout/cart.php
@@ -93,11 +93,11 @@ class Cart extends \Opencart\System\Engine\Model {
 	 *
 	 * @param array<int, array<string, mixed>> $totals
 	 * @param array<int, float>                $taxes
-	 * @param int                              $total
+	 * @param float                            $total
 	 *
 	 * @return void
 	 */
-	public function getTotals(array &$totals, array &$taxes, int &$total): void {
+	public function getTotals(array &$totals, array &$taxes, float &$total): void {
 		$sort_order = [];
 
 		// Extensions
